feat(ui): edit monitor physical size from Settings → Monitors
Adds an "Edit size" button on each monitor row that pops open an inline
editor: diagonal-inches input as primary, with live width/height in mm
the user can also override directly. Saves via a new
`set_monitor_physical_size` IPC method backed by `write_monitor_block`,
which preserves comments in the user's config. After saving, the toast
offers a "Re-apply now" action when a profile is currently active so
bezel math picks up the new size without a manual re-apply.

Also migrates `physical_mm` storage from `[u32; 2]` to `[f64; 2]` end
to end (config / runtime / IPC / layout) so the UI's 1dp precision
round-trips cleanly. Existing TOML configs with integer literals
(`physical_mm = [597, 336]`) continue to parse via serde's int→float
coercion; new writes use float form.

+++ crates/superpanels-core/src/config/merge.rs
@@ -33,13 +33,13 @@ fn matches_block(block: &MonitorConfig, m: &Monitor) -> bool {
     false
 }
 
-fn compute_ppi(resolution: (u32, u32), physical_mm: (u32, u32), rotation: Rotation) -> f64 {
+fn compute_ppi(resolution: (u32, u32), physical_mm: (f64, f64), rotation: Rotation) -> f64 {
     let (px_w, px_h) = match rotation {
         Rotation::None | Rotation::Inverted => resolution,
         Rotation::Right | Rotation::Left => (resolution.1, resolution.0),
     };
     let (mm_w, mm_h) = physical_mm;
-    let ppi_w = f64::from(px_w) / (f64::from(mm_w) / MM_PER_INCH);
-    let ppi_h = f64::from(px_h) / (f64::from(mm_h) / MM_PER_INCH);
+    let ppi_w = f64::from(px_w) / (mm_w / MM_PER_INCH);
+    let ppi_h = f64::from(px_h) / (mm_h / MM_PER_INCH);
     f64::midpoint(ppi_w, ppi_h)
 }
